requirements.txt
Flask
psycopg2-binary
requests
python-dotenv
pandas
sqlalchemy